+++ src/Oxrun/Command/Oxid/ShopListCommand.php
@@ -11,6 +11,7 @@ namespace Oxrun\Command\Oxid;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
@@ -23,6 +24,7 @@ class ShopListCommand extends Command
     protected function configure()
     {
         $this->setName('oxid:shops')
+            ->addOption('only-ids', 'i', InputOption::VALUE_NONE, 'show only Shop id\'s. eg. "oe-console oxid:shops --only-ids | xargs -tn1 oe-console ... --shop-id "')
             ->setDescription('Lists the shops');
     }
 
@@ -42,6 +44,10 @@ class ShopListCommand extends Command
                      $output->getVerbosity() & OutputInterface::VERBOSITY_VERY_VERBOSE ||
                      $output->getVerbosity() & OutputInterface::VERBOSITY_DEBUG;
 
+        if ($input->getOption('only-ids')) {
+            $this->displayShopIds($output, $oxShopList);
+            return 0;
+        }
         if ($isVerbose) {
             $this->displayVerbose($oxShopList, $table);
         } else {
@@ -50,6 +56,17 @@ class ShopListCommand extends Command
 
         $table->render();
         return 0;
+    }
+
+    /**
+     * @param OutputInterface $output
+     * @param \OxidEsales\Eshop\Application\Model\ShopList $oxShopList
+     */
+    protected function displayShopIds(OutputInterface $output, \OxidEsales\Eshop\Application\Model\ShopList $oxShopList)
+    {
+        foreach ($oxShopList as $oxShop) {
+            $output->writeln($oxShop->getId());
+        }
     }
 
     /**